refactor: 익명클래스 적용
