--- src/main/java/springbook/user/dao/AddStatement.java
@@ -1,29 +1,0 @@
-package springbook.user.dao;
-
-import springbook.user.domain.User;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-/**
- * @author Kj Nam
- * @since 2017-05-18
- */
-public class AddStatement implements StatementStrategy {
-    private User user;
-
-    public AddStatement(User user) {
-        this.user = user;
-    }
-
-    @Override
-    public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
-        PreparedStatement ps = c.prepareStatement("INSERT INTO USERS(id, name, password) VALUES(?,?,?)");
-        ps.setString(1, user.getId());
-        ps.setString(2, user.getName());
-        ps.setString(3, user.getPassword());
-
-        return ps;
-    }
-}
